LayoutTests/platform/chromium/fast/dom/wrapper-classes-expected.txt

This tests wrapper class names for JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Basics

PASS jsWrapperClass(document.createAttribute()) is 'Attr'
FAIL jsWrapperClass(document.createAttribute().__proto__) should be AttrPrototype. Was Object.
FAIL jsWrapperClass(document.createAttribute().constructor) should be AttrConstructor. Was Function.
PASS jsWrapperClass(document.createComment()) is 'Comment'
FAIL jsWrapperClass(document.createComment().__proto__) should be CommentPrototype. Was Object.
FAIL jsWrapperClass(document.createComment().constructor) should be CommentConstructor. Was Function.
PASS jsWrapperClass(document.createDocumentFragment()) is 'DocumentFragment'
FAIL jsWrapperClass(document.createDocumentFragment().__proto__) should be DocumentFragmentPrototype. Was Object.
FAIL jsWrapperClass(document.createDocumentFragment().constructor) should be DocumentFragmentConstructor. Was Function.
PASS jsWrapperClass(document.implementation) is 'DOMImplementation'
FAIL jsWrapperClass(document.implementation.__proto__) should be DOMImplementationPrototype. Was Object.
FAIL jsWrapperClass(document.implementation.constructor) should be DOMImplementationConstructor. Was Function.
PASS jsWrapperClass(root.attributes) is 'NamedNodeMap'
FAIL jsWrapperClass(root.attributes.__proto__) should be NamedNodeMapPrototype. Was Object.
FAIL jsWrapperClass(root.attributes.constructor) should be NamedNodeMapConstructor. Was Function.
PASS jsWrapperClass(document.createNodeIterator(root, 0, null, false)) is 'NodeIterator'
FAIL jsWrapperClass(document.createNodeIterator(root, 0, null, false).__proto__) should be NodeIteratorPrototype. Was Object.
FAIL jsWrapperClass(document.createNodeIterator(root, 0, null, false).constructor) should be NodeIteratorConstructor. Was Function.
PASS jsWrapperClass(document.getElementsByTagName('div')) is 'NodeList'
FAIL jsWrapperClass(document.getElementsByTagName('div').__proto__) should be NodeListPrototype. Was Object.
FAIL jsWrapperClass(document.getElementsByTagName('div').constructor) should be NodeListConstructor. Was Function.
PASS jsWrapperClass(document.createRange()) is 'Range'
FAIL jsWrapperClass(document.createRange().__proto__) should be RangePrototype. Was Object.
FAIL jsWrapperClass(document.createRange().constructor) should be RangeConstructor. Was Function.
PASS jsWrapperClass(document.createTextNode()) is 'Text'
FAIL jsWrapperClass(document.createTextNode().__proto__) should be TextPrototype. Was Object.
FAIL jsWrapperClass(document.createTextNode().constructor) should be TextConstructor. Was Function.
PASS jsWrapperClass(document.createTreeWalker(root, 0, null, false)) is 'TreeWalker'
FAIL jsWrapperClass(document.createTreeWalker(root, 0, null, false).__proto__) should be TreeWalkerPrototype. Was Object.
FAIL jsWrapperClass(document.createTreeWalker(root, 0, null, false).constructor) should be TreeWalkerConstructor. Was Function.

XML

PASS jsWrapperClass(xmlDocument) is 'Document'
FAIL jsWrapperClass(xmlDocument.__proto__) should be DocumentPrototype. Was Object.
FAIL jsWrapperClass(xmlDocument.constructor) should be DocumentConstructor. Was Function.
PASS jsWrapperClass(xmlDocument.createCDATASection()) is 'CDATASection'
FAIL jsWrapperClass(xmlDocument.createCDATASection().__proto__) should be CDATASectionPrototype. Was Object.
FAIL jsWrapperClass(xmlDocument.createCDATASection().constructor) should be CDATASectionConstructor. Was Function.
PASS jsWrapperClass(xmlDocument.createElementNS('x', 'x', 'x')) is 'Element'
FAIL jsWrapperClass(xmlDocument.createElementNS('x', 'x', 'x').__proto__) should be ElementPrototype. Was Object.
FAIL jsWrapperClass(xmlDocument.createElementNS('x', 'x', 'x').constructor) should be ElementConstructor. Was Function.
PASS jsWrapperClass(xmlDocument.createEntityReference()) is 'EntityReference'
FAIL jsWrapperClass(xmlDocument.createEntityReference().__proto__) should be EntityReferencePrototype. Was Object.
FAIL jsWrapperClass(xmlDocument.createEntityReference().constructor) should be EntityReferenceConstructor. Was Function.
PASS jsWrapperClass(xmlDocument.createProcessingInstruction()) is 'ProcessingInstruction'
FAIL jsWrapperClass(xmlDocument.createProcessingInstruction().__proto__) should be ProcessingInstructionPrototype. Was Object.
FAIL jsWrapperClass(xmlDocument.createProcessingInstruction().constructor) should be ProcessingInstructionConstructor. Was Function.

Events

PASS jsWrapperClass(document.createEvent('Event')) is 'Event'
FAIL jsWrapperClass(document.createEvent('Event').__proto__) should be EventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('Event').constructor) should be EventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('KeyboardEvent')) is 'KeyboardEvent'
FAIL jsWrapperClass(document.createEvent('KeyboardEvent').__proto__) should be KeyboardEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('KeyboardEvent').constructor) should be KeyboardEventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('MouseEvent')) is 'MouseEvent'
FAIL jsWrapperClass(document.createEvent('MouseEvent').__proto__) should be MouseEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('MouseEvent').constructor) should be MouseEventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('MutationEvent')) is 'MutationEvent'
FAIL jsWrapperClass(document.createEvent('MutationEvent').__proto__) should be MutationEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('MutationEvent').constructor) should be MutationEventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('OverflowEvent')) is 'OverflowEvent'
FAIL jsWrapperClass(document.createEvent('OverflowEvent').__proto__) should be OverflowEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('OverflowEvent').constructor) should be OverflowEventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('TextEvent')) is 'TextEvent'
FAIL jsWrapperClass(document.createEvent('TextEvent').__proto__) should be TextEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('TextEvent').constructor) should be TextEventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('UIEvent')) is 'UIEvent'
FAIL jsWrapperClass(document.createEvent('UIEvent').__proto__) should be UIEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('UIEvent').constructor) should be UIEventConstructor. Was Function.
PASS jsWrapperClass(document.createEvent('WheelEvent')) is 'WheelEvent'
FAIL jsWrapperClass(document.createEvent('WheelEvent').__proto__) should be WheelEventPrototype. Was Object.
FAIL jsWrapperClass(document.createEvent('WheelEvent').constructor) should be WheelEventConstructor. Was Function.

CSS DOM

PASS jsWrapperClass(document.styleSheets) is 'StyleSheetList'
FAIL jsWrapperClass(document.styleSheets.__proto__) should be StyleSheetListPrototype. Was Object.
FAIL jsWrapperClass(document.styleSheets.constructor) should be StyleSheetListConstructor. Was Function.
PASS jsWrapperClass(stylesheet) is 'CSSStyleSheet'
FAIL jsWrapperClass(stylesheet.__proto__) should be CSSStyleSheetPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.constructor) should be CSSStyleSheetConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules) is 'CSSRuleList'
FAIL jsWrapperClass(stylesheet.cssRules.__proto__) should be CSSRuleListPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.constructor) should be CSSRuleListConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0)) is 'CSSStyleRule'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).__proto__) should be CSSStyleRulePrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).constructor) should be CSSStyleRuleConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0).style) is 'CSSStyleDeclaration'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.__proto__) should be CSSStyleDeclarationPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.constructor) should be CSSStyleDeclarationConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color')) is 'CSSPrimitiveValue'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').__proto__) should be CSSPrimitiveValuePrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').constructor) should be CSSPrimitiveValueConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').getRGBColorValue()) is 'RGBColor'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').getRGBColorValue().__proto__) should be RGBColorPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').getRGBColorValue().constructor) should be RGBColorConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('clip').getRectValue()) is 'Rect'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('clip').getRectValue().__proto__) should be RectPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('clip').getRectValue().constructor) should be RectConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content')) is 'CSSValueList'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').__proto__) should be CSSValueListPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').constructor) should be CSSValueListConstructor. Was Function.
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').item(0).getCounterValue()) is 'Counter'
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').item(0).getCounterValue().__proto__) should be CounterPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').item(0).getCounterValue().constructor) should be CounterConstructor. Was Function.
PASS jsWrapperClass(stylesheet.media) is 'MediaList'
FAIL jsWrapperClass(stylesheet.media.__proto__) should be MediaListPrototype. Was Object.
FAIL jsWrapperClass(stylesheet.media.constructor) should be MediaListConstructor. Was Function.

XPath

PASS jsWrapperClass(document.createExpression('//*', document.createNSResolver(document))) is 'XPathExpression'
FAIL jsWrapperClass(document.createExpression('//*', document.createNSResolver(document)).__proto__) should be XPathExpressionPrototype. Was Object.
FAIL jsWrapperClass(document.createExpression('//*', document.createNSResolver(document)).constructor) should be XPathExpressionConstructor. Was Function.
PASS jsWrapperClass(document.evaluate('//*', document, document.createNSResolver(document), 0, null)) is 'XPathResult'
FAIL jsWrapperClass(document.evaluate('//*', document, document.createNSResolver(document), 0, null).__proto__) should be XPathResultPrototype. Was Object.
FAIL jsWrapperClass(document.evaluate('//*', document, document.createNSResolver(document), 0, null).constructor) should be XPathResultConstructor. Was Function.

Other

FAIL jsWrapperClass(window) should be DOMWindow. Was global.
FAIL jsWrapperClass(window.__proto__) should be DOMWindowPrototype. Was Object.
FAIL jsWrapperClass(window.constructor) should be DOMWindowConstructor. Was Function.

HTML DOM

PASS jsWrapperClass(root.children) is 'HTMLCollection'
FAIL jsWrapperClass(root.children.__proto__) should be HTMLCollectionPrototype. Was Object.
FAIL jsWrapperClass(root.children.constructor) should be HTMLCollectionConstructor. Was Function.
PASS jsWrapperClass(document) is 'HTMLDocument'
FAIL jsWrapperClass(document.__proto__) should be HTMLDocumentPrototype. Was Object.
FAIL jsWrapperClass(document.constructor) should be HTMLDocumentConstructor. Was Function.
PASS jsWrapperClass(document.createElement('select').options) is 'HTMLOptionsCollection'
FAIL jsWrapperClass(document.createElement('select').options.__proto__) should be HTMLOptionsCollectionPrototype. Was Object.
FAIL jsWrapperClass(document.createElement('select').options.constructor) should be HTMLOptionsCollectionConstructor. Was Function.

HTML Elements

PASS tagJSWrapperClass('a') is 'HTMLAnchorElement'
FAIL tagJSWrapperPrototypeClass('a') should be HTMLAnchorElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('a') should be HTMLAnchorElementConstructor. Was Function.
PASS tagJSWrapperClass('abbr') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('abbr') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('abbr') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('acronym') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('acronym') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('acronym') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('address') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('address') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('address') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('applet') is 'HTMLAppletElement'
FAIL tagJSWrapperPrototypeClass('applet') should be HTMLAppletElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('applet') should be HTMLAppletElementConstructor. Was Function.
PASS tagJSWrapperClass('area') is 'HTMLAreaElement'
FAIL tagJSWrapperPrototypeClass('area') should be HTMLAreaElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('area') should be HTMLAreaElementConstructor. Was Function.
PASS tagJSWrapperClass('article') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('article') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('article') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('aside') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('aside') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('aside') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('b') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('b') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('b') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('base') is 'HTMLBaseElement'
FAIL tagJSWrapperPrototypeClass('base') should be HTMLBaseElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('base') should be HTMLBaseElementConstructor. Was Function.
PASS tagJSWrapperClass('basefont') is 'HTMLBaseFontElement'
FAIL tagJSWrapperPrototypeClass('basefont') should be HTMLBaseFontElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('basefont') should be HTMLBaseFontElementConstructor. Was Function.
PASS tagJSWrapperClass('bdo') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('bdo') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('bdo') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('big') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('big') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('big') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('blockquote') is 'HTMLQuoteElement'
FAIL tagJSWrapperPrototypeClass('blockquote') should be HTMLQuoteElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('blockquote') should be HTMLQuoteElementConstructor. Was Function.
PASS tagJSWrapperClass('body') is 'HTMLBodyElement'
FAIL tagJSWrapperPrototypeClass('body') should be HTMLBodyElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('body') should be HTMLBodyElementConstructor. Was Function.
PASS tagJSWrapperClass('br') is 'HTMLBRElement'
FAIL tagJSWrapperPrototypeClass('br') should be HTMLBRElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('br') should be HTMLBRElementConstructor. Was Function.
PASS tagJSWrapperClass('button') is 'HTMLButtonElement'
FAIL tagJSWrapperPrototypeClass('button') should be HTMLButtonElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('button') should be HTMLButtonElementConstructor. Was Function.
PASS tagJSWrapperClass('canvas') is 'HTMLCanvasElement'
FAIL tagJSWrapperPrototypeClass('canvas') should be HTMLCanvasElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('canvas') should be HTMLCanvasElementConstructor. Was Function.
PASS tagJSWrapperClass('caption') is 'HTMLTableCaptionElement'
FAIL tagJSWrapperPrototypeClass('caption') should be HTMLTableCaptionElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('caption') should be HTMLTableCaptionElementConstructor. Was Function.
PASS tagJSWrapperClass('center') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('center') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('center') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('cite') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('cite') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('cite') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('code') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('code') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('code') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('col') is 'HTMLTableColElement'
FAIL tagJSWrapperPrototypeClass('col') should be HTMLTableColElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('col') should be HTMLTableColElementConstructor. Was Function.
PASS tagJSWrapperClass('colgroup') is 'HTMLTableColElement'
FAIL tagJSWrapperPrototypeClass('colgroup') should be HTMLTableColElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('colgroup') should be HTMLTableColElementConstructor. Was Function.
PASS tagJSWrapperClass('dd') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('dd') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('dd') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('del') is 'HTMLModElement'
FAIL tagJSWrapperPrototypeClass('del') should be HTMLModElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('del') should be HTMLModElementConstructor. Was Function.
PASS tagJSWrapperClass('dfn') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('dfn') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('dfn') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('dir') is 'HTMLDirectoryElement'
FAIL tagJSWrapperPrototypeClass('dir') should be HTMLDirectoryElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('dir') should be HTMLDirectoryElementConstructor. Was Function.
PASS tagJSWrapperClass('div') is 'HTMLDivElement'
FAIL tagJSWrapperPrototypeClass('div') should be HTMLDivElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('div') should be HTMLDivElementConstructor. Was Function.
PASS tagJSWrapperClass('dl') is 'HTMLDListElement'
FAIL tagJSWrapperPrototypeClass('dl') should be HTMLDListElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('dl') should be HTMLDListElementConstructor. Was Function.
PASS tagJSWrapperClass('dt') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('dt') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('dt') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('em') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('em') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('em') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('embed') is 'HTMLEmbedElement'
FAIL tagJSWrapperPrototypeClass('embed') should be HTMLEmbedElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('embed') should be HTMLEmbedElementConstructor. Was Function.
PASS tagJSWrapperClass('fieldset') is 'HTMLFieldSetElement'
FAIL tagJSWrapperPrototypeClass('fieldset') should be HTMLFieldSetElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('fieldset') should be HTMLFieldSetElementConstructor. Was Function.
PASS tagJSWrapperClass('font') is 'HTMLFontElement'
FAIL tagJSWrapperPrototypeClass('font') should be HTMLFontElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('font') should be HTMLFontElementConstructor. Was Function.
PASS tagJSWrapperClass('footer') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('footer') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('footer') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('form') is 'HTMLFormElement'
FAIL tagJSWrapperPrototypeClass('form') should be HTMLFormElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('form') should be HTMLFormElementConstructor. Was Function.
PASS tagJSWrapperClass('frame') is 'HTMLFrameElement'
FAIL tagJSWrapperPrototypeClass('frame') should be HTMLFrameElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('frame') should be HTMLFrameElementConstructor. Was Function.
PASS tagJSWrapperClass('frameset') is 'HTMLFrameSetElement'
FAIL tagJSWrapperPrototypeClass('frameset') should be HTMLFrameSetElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('frameset') should be HTMLFrameSetElementConstructor. Was Function.
PASS tagJSWrapperClass('head') is 'HTMLHeadElement'
FAIL tagJSWrapperPrototypeClass('head') should be HTMLHeadElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('head') should be HTMLHeadElementConstructor. Was Function.
PASS tagJSWrapperClass('h1') is 'HTMLHeadingElement'
FAIL tagJSWrapperPrototypeClass('h1') should be HTMLHeadingElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('h1') should be HTMLHeadingElementConstructor. Was Function.
PASS tagJSWrapperClass('h2') is 'HTMLHeadingElement'
FAIL tagJSWrapperPrototypeClass('h2') should be HTMLHeadingElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('h2') should be HTMLHeadingElementConstructor. Was Function.
PASS tagJSWrapperClass('h3') is 'HTMLHeadingElement'
FAIL tagJSWrapperPrototypeClass('h3') should be HTMLHeadingElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('h3') should be HTMLHeadingElementConstructor. Was Function.
PASS tagJSWrapperClass('h4') is 'HTMLHeadingElement'
FAIL tagJSWrapperPrototypeClass('h4') should be HTMLHeadingElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('h4') should be HTMLHeadingElementConstructor. Was Function.
PASS tagJSWrapperClass('h5') is 'HTMLHeadingElement'
FAIL tagJSWrapperPrototypeClass('h5') should be HTMLHeadingElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('h5') should be HTMLHeadingElementConstructor. Was Function.
PASS tagJSWrapperClass('h6') is 'HTMLHeadingElement'
FAIL tagJSWrapperPrototypeClass('h6') should be HTMLHeadingElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('h6') should be HTMLHeadingElementConstructor. Was Function.
PASS tagJSWrapperClass('head') is 'HTMLHeadElement'
FAIL tagJSWrapperPrototypeClass('head') should be HTMLHeadElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('head') should be HTMLHeadElementConstructor. Was Function.
PASS tagJSWrapperClass('header') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('header') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('header') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('hgroup') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('hgroup') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('hgroup') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('hr') is 'HTMLHRElement'
FAIL tagJSWrapperPrototypeClass('hr') should be HTMLHRElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('hr') should be HTMLHRElementConstructor. Was Function.
PASS tagJSWrapperClass('html') is 'HTMLHtmlElement'
FAIL tagJSWrapperPrototypeClass('html') should be HTMLHtmlElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('html') should be HTMLHtmlElementConstructor. Was Function.
PASS tagJSWrapperClass('i') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('i') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('i') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('iframe') is 'HTMLIFrameElement'
FAIL tagJSWrapperPrototypeClass('iframe') should be HTMLIFrameElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('iframe') should be HTMLIFrameElementConstructor. Was Function.
PASS tagJSWrapperClass('image') is 'HTMLImageElement'
FAIL tagJSWrapperPrototypeClass('image') should be HTMLImageElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('image') should be HTMLImageElementConstructor. Was Function.
PASS tagJSWrapperClass('img') is 'HTMLImageElement'
FAIL tagJSWrapperPrototypeClass('img') should be HTMLImageElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('img') should be HTMLImageElementConstructor. Was Function.
PASS tagJSWrapperClass('input') is 'HTMLInputElement'
FAIL tagJSWrapperPrototypeClass('input') should be HTMLInputElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('input') should be HTMLInputElementConstructor. Was Function.
PASS tagJSWrapperClass('ins') is 'HTMLModElement'
FAIL tagJSWrapperPrototypeClass('ins') should be HTMLModElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('ins') should be HTMLModElementConstructor. Was Function.
PASS tagJSWrapperClass('kbd') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('kbd') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('kbd') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('keygen') is 'HTMLKeygenElement'
FAIL tagJSWrapperPrototypeClass('keygen') should be HTMLKeygenElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('keygen') should be HTMLKeygenElementConstructor. Was Function.
PASS tagJSWrapperClass('label') is 'HTMLLabelElement'
FAIL tagJSWrapperPrototypeClass('label') should be HTMLLabelElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('label') should be HTMLLabelElementConstructor. Was Function.
PASS tagJSWrapperClass('layer') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('layer') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('layer') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('legend') is 'HTMLLegendElement'
FAIL tagJSWrapperPrototypeClass('legend') should be HTMLLegendElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('legend') should be HTMLLegendElementConstructor. Was Function.
PASS tagJSWrapperClass('li') is 'HTMLLIElement'
FAIL tagJSWrapperPrototypeClass('li') should be HTMLLIElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('li') should be HTMLLIElementConstructor. Was Function.
PASS tagJSWrapperClass('link') is 'HTMLLinkElement'
FAIL tagJSWrapperPrototypeClass('link') should be HTMLLinkElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('link') should be HTMLLinkElementConstructor. Was Function.
PASS tagJSWrapperClass('listing') is 'HTMLPreElement'
FAIL tagJSWrapperPrototypeClass('listing') should be HTMLPreElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('listing') should be HTMLPreElementConstructor. Was Function.
PASS tagJSWrapperClass('map') is 'HTMLMapElement'
FAIL tagJSWrapperPrototypeClass('map') should be HTMLMapElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('map') should be HTMLMapElementConstructor. Was Function.
PASS tagJSWrapperClass('marquee') is 'HTMLMarqueeElement'
FAIL tagJSWrapperPrototypeClass('marquee') should be HTMLMarqueeElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('marquee') should be HTMLMarqueeElementConstructor. Was Function.
PASS tagJSWrapperClass('menu') is 'HTMLMenuElement'
FAIL tagJSWrapperPrototypeClass('menu') should be HTMLMenuElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('menu') should be HTMLMenuElementConstructor. Was Function.
PASS tagJSWrapperClass('meta') is 'HTMLMetaElement'
FAIL tagJSWrapperPrototypeClass('meta') should be HTMLMetaElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('meta') should be HTMLMetaElementConstructor. Was Function.
PASS tagJSWrapperClass('nav') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('nav') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('nav') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('nobr') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('nobr') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('nobr') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('noembed') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('noembed') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('noembed') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('noframes') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('noframes') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('noframes') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('nolayer') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('nolayer') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('nolayer') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('noscript') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('noscript') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('noscript') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('object') is 'HTMLObjectElement'
FAIL tagJSWrapperPrototypeClass('object') should be HTMLObjectElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('object') should be HTMLObjectElementConstructor. Was Function.
PASS tagJSWrapperClass('ol') is 'HTMLOListElement'
FAIL tagJSWrapperPrototypeClass('ol') should be HTMLOListElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('ol') should be HTMLOListElementConstructor. Was Function.
PASS tagJSWrapperClass('optgroup') is 'HTMLOptGroupElement'
FAIL tagJSWrapperPrototypeClass('optgroup') should be HTMLOptGroupElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('optgroup') should be HTMLOptGroupElementConstructor. Was Function.
PASS tagJSWrapperClass('option') is 'HTMLOptionElement'
FAIL tagJSWrapperPrototypeClass('option') should be HTMLOptionElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('option') should be HTMLOptionElementConstructor. Was Function.
PASS tagJSWrapperClass('p') is 'HTMLParagraphElement'
FAIL tagJSWrapperPrototypeClass('p') should be HTMLParagraphElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('p') should be HTMLParagraphElementConstructor. Was Function.
PASS tagJSWrapperClass('param') is 'HTMLParamElement'
FAIL tagJSWrapperPrototypeClass('param') should be HTMLParamElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('param') should be HTMLParamElementConstructor. Was Function.
PASS tagJSWrapperClass('plaintext') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('plaintext') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('plaintext') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('pre') is 'HTMLPreElement'
FAIL tagJSWrapperPrototypeClass('pre') should be HTMLPreElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('pre') should be HTMLPreElementConstructor. Was Function.
PASS tagJSWrapperClass('q') is 'HTMLQuoteElement'
FAIL tagJSWrapperPrototypeClass('q') should be HTMLQuoteElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('q') should be HTMLQuoteElementConstructor. Was Function.
PASS tagJSWrapperClass('rp') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('rp') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('rp') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('rt') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('rt') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('rt') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('ruby') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('ruby') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('ruby') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('s') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('s') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('s') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('samp') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('samp') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('samp') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('script') is 'HTMLScriptElement'
FAIL tagJSWrapperPrototypeClass('script') should be HTMLScriptElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('script') should be HTMLScriptElementConstructor. Was Function.
PASS tagJSWrapperClass('section') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('section') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('section') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('select') is 'HTMLSelectElement'
FAIL tagJSWrapperPrototypeClass('select') should be HTMLSelectElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('select') should be HTMLSelectElementConstructor. Was Function.
PASS tagJSWrapperClass('small') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('small') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('small') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('span') is 'HTMLSpanElement'
FAIL tagJSWrapperPrototypeClass('span') should be HTMLSpanElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('span') should be HTMLSpanElementConstructor. Was Function.
PASS tagJSWrapperClass('strike') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('strike') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('strike') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('strong') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('strong') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('strong') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('style') is 'HTMLStyleElement'
FAIL tagJSWrapperPrototypeClass('style') should be HTMLStyleElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('style') should be HTMLStyleElementConstructor. Was Function.
PASS tagJSWrapperClass('sub') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('sub') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('sub') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('sup') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('sup') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('sup') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('table') is 'HTMLTableElement'
FAIL tagJSWrapperPrototypeClass('table') should be HTMLTableElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('table') should be HTMLTableElementConstructor. Was Function.
PASS tagJSWrapperClass('tbody') is 'HTMLTableSectionElement'
FAIL tagJSWrapperPrototypeClass('tbody') should be HTMLTableSectionElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('tbody') should be HTMLTableSectionElementConstructor. Was Function.
PASS tagJSWrapperClass('td') is 'HTMLTableCellElement'
FAIL tagJSWrapperPrototypeClass('td') should be HTMLTableCellElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('td') should be HTMLTableCellElementConstructor. Was Function.
PASS tagJSWrapperClass('textarea') is 'HTMLTextAreaElement'
FAIL tagJSWrapperPrototypeClass('textarea') should be HTMLTextAreaElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('textarea') should be HTMLTextAreaElementConstructor. Was Function.
PASS tagJSWrapperClass('tfoot') is 'HTMLTableSectionElement'
FAIL tagJSWrapperPrototypeClass('tfoot') should be HTMLTableSectionElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('tfoot') should be HTMLTableSectionElementConstructor. Was Function.
PASS tagJSWrapperClass('th') is 'HTMLTableCellElement'
FAIL tagJSWrapperPrototypeClass('th') should be HTMLTableCellElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('th') should be HTMLTableCellElementConstructor. Was Function.
PASS tagJSWrapperClass('thead') is 'HTMLTableSectionElement'
FAIL tagJSWrapperPrototypeClass('thead') should be HTMLTableSectionElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('thead') should be HTMLTableSectionElementConstructor. Was Function.
PASS tagJSWrapperClass('title') is 'HTMLTitleElement'
FAIL tagJSWrapperPrototypeClass('title') should be HTMLTitleElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('title') should be HTMLTitleElementConstructor. Was Function.
PASS tagJSWrapperClass('tr') is 'HTMLTableRowElement'
FAIL tagJSWrapperPrototypeClass('tr') should be HTMLTableRowElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('tr') should be HTMLTableRowElementConstructor. Was Function.
PASS tagJSWrapperClass('tt') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('tt') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('tt') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('u') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('u') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('u') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('ul') is 'HTMLUListElement'
FAIL tagJSWrapperPrototypeClass('ul') should be HTMLUListElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('ul') should be HTMLUListElementConstructor. Was Function.
PASS tagJSWrapperClass('var') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('var') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('var') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('wbr') is 'HTMLElement'
FAIL tagJSWrapperPrototypeClass('wbr') should be HTMLElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('wbr') should be HTMLElementConstructor. Was Function.
PASS tagJSWrapperClass('xmp') is 'HTMLPreElement'
FAIL tagJSWrapperPrototypeClass('xmp') should be HTMLPreElementPrototype. Was Object.
FAIL tagJSWrapperConstructorClass('xmp') should be HTMLPreElementConstructor. Was Function.
